requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
accelerate=0.29.2=pyhd8ed1ab_0
aiohttp=3.9.3=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
annotated-types=0.6.0=py39h06a4308_0
aom=3.6.0=h6a678d5_0
arch-py=6.3.0=py39h44dd56e_0
async-timeout=4.0.3=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
axial-positional-embedding=0.2.1=pyhd8ed1ab_0
black=24.8.0=py39h06a4308_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
bottleneck=1.3.7=py39ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py39h6a678d5_7
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.11.26=h06a4308_0
calflops=0.3.2=pypi_0
certifi=2024.8.30=py39h06a4308_0
cfitsio=3.470=h5893167_7
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.2.0=py39hdb19cb5_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=0
cuda-runtime=12.1.0=0
cuda-version=11.8=hcce14f8_3
cudatoolkit=11.8.0=h6a678d5_0
cudnn=8.9.7.29=hbc23b4c_3
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dav1d=1.2.1=h5eee18b_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd8ed1ab_0
deepspeed=0.14.0=cuda118py39h87c11cd_200
einops=0.7.0=pyhd8ed1ab_1
et_xmlfile=1.1.0=py39h06a4308_0
expat=2.6.2=h6a678d5_0
fastdtw=0.3.4=py39h8003fee_4
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.51.0=py39h5eee18b_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py39h5eee18b_0
fsspec=2024.3.1=pyhca7485f_0
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
hjson-py=3.1.0=pyhd8ed1ab_0
huggingface_hub=0.23.4=pyhd8ed1ab_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
imagecodecs=2023.1.23=py39hc4b7b5f_0
imageio=2.33.1=py39h06a4308_0
importlib-metadata=7.0.1=py39h06a4308_0
importlib_resources=6.1.1=py39h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy_loader=0.4=pyhd8ed1ab_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libaec=1.0.4=he6710b0_1
libaio=0.3.113=h5eee18b_0
libavif=0.11.1=h5eee18b_0
libblas=3.9.0=20_linux64_mkl
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=20_linux64_mkl
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.5.147=0
libcurl=8.9.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=14.2.0=h77fa898_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=20_linux64_mkl
libllvm14=14.0.6=hdb19cb5_3
libmagma=2.7.2=h09b5827_2
libmagma_sparse=2.7.2=h09b5827_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=12.17=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=13.2.0=hc0a3c3a_6
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libtorch=2.1.2=cuda118_h9f31b6d_304
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=hd590300_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
libxslt=1.1.41=h097e994_0
libzlib=1.2.13=hd590300_5
libzopfli=1.0.3=he6710b0_0
llvm-openmp=18.1.5=ha31de31_0
local-attention=1.9.3=pyhd8ed1ab_0
lxml=5.2.1=py39h57af460_1
lz4-c=1.9.4=h6a678d5_0
magma=2.7.2=h4aca40b_3
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.8.4=py39h06a4308_0
matplotlib-base=3.8.4=py39h1128e8f_0
mkl=2023.2.0=h84fe81f_50496
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mne-base=1.7.1=pyha770c72_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.4=py39h5eee18b_0
mypy_extensions=1.0.0=py39h06a4308_0
mysql=5.7.24=h721c034_2
nccl=2.21.5.1=h6103f9b_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.1.5=py39h5eee18b_0
openssl=3.4.0=hb9d3cd8_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=py39h6a678d5_0
pathspec=0.10.3=py39h06a4308_0
patool=1.15.0=pyhd8ed1ab_0
patsy=0.5.3=py39h06a4308_0
pcre2=10.42=hebb0a14_0
performer-pytorch=1.1.4=pyhd8ed1ab_0
pillow=10.2.0=py39h5eee18b_0
pip=24.2=py39h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
ply=3.11=py39h06a4308_0
pooch=1.8.2=pyhd8ed1ab_0
protobuf=4.25.3=py39h12ddb61_0
psutil=5.9.8=py39hd1e30aa_0
py-cpuinfo=9.0.0=py39h06a4308_0
pydantic=2.5.3=py39h06a4308_0
pydantic-core=2.14.6=py39hb02cf49_0
pyg=2.5.2=py39_torch_2.1.0_cu121
pynvml=11.5.0=pyhd8ed1ab_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=2.1.2=cuda118_py39hd44be3b_304
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyyaml=6.0.1=py39h5eee18b_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
regex=2023.10.3=py39h5eee18b_0
requests=2.31.0=py39h06a4308_1
safetensors=0.4.3=py39h9fdd4d6_0
scikit-base=0.8.2=pyhecae5ae_0
scikit-image=0.24.0=py39h1128e8f_0
scikit-learn=1.3.0=py39h1128e8f_1
scipy=1.12.0=py39h5f9d8c6_0
seaborn=0.13.2=py39h06a4308_0
selenium=3.141.0=py39h27cfd23_1000
sentencepiece=0.1.99=py39hdb19cb5_0
setuptools=68.2.2=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sktime=0.31.0=py39hf3d152e_0
sleef=3.5.1=h9b69904_2
snappy=1.2.1=h6a678d5_0
sqlite=3.41.2=h5eee18b_0
statsmodels=0.14.0=py39ha9d4c09_0
sympy=1.12=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.4.12=py39h06a4308_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=py39h22610ee_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py39h06a4308_0
torchaudio=2.1.2=py39_cu121
torchinfo=1.8.0=pyhd8ed1ab_0
torchtriton=2.2.0=py39
torchvision=0.15.2=cuda118py39h196c800_0
tornado=6.3.3=py39h5eee18b_0
tqdm=4.66.2=pyhd8ed1ab_0
transformers=4.37.2=py39h06a4308_0
typing-extensions=4.9.0=py39h06a4308_1
typing_extensions=4.9.0=py39h06a4308_1
tzdata=2024a=h04d1e81_0
unicodedata2=15.1.0=py39h5eee18b_0
urllib3=2.0.3=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py39h5eee18b_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=py39h06a4308_0
zlib=1.2.13=hd590300_5
zstd=1.5.6=ha6fb4c9_0
